Modularize dataset utilities and clean code
