+++ Classificazione/ViT/transformer_model_temperatura.py
