+++ Segmentazione/CNN/model.py
